Refactor

See merge request codecool/application-system-admin!43

+++ backend/src/main/java/com/codecool/appsystem/admin/model/dto/ApplicantInfoDTO.java
@@ -15,7 +15,7 @@ public class ApplicantInfoDTO extends RestResponseDTO {
     private String name;
     private String location;
     private String status;
-    private Long attempts;
+    private Integer attempts;
     private Boolean blacklisted;
     private Date processStartedAt;
     private String email;
